allow opportunity for hooks when setting up express app
